usecases/conversation-coach/artifacts/cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cc-coach"
version = "0.1.0"
description = "Conversation Coach CLI - Local orchestration for CCAI Insights pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Conversation Coach Team" }
]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-cloud-contact-center-insights>=1.15.0",
    "google-cloud-bigquery>=3.13.0",
    "google-cloud-storage>=2.14.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
cc-coach = "cc_coach.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["cc_coach"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=cc_coach --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
